pyproject.toml
[project]
name = "fair-mast"
version = "1.0"
description = "A data management system for Tokamak data"
readme = "README.md"
requires-python = ">=3.9, <3.13"
authors = [
]
maintainers = [
    {name = "Samuel Jackson", email = "samuel.jackson@ukaea.uk"},
    {name = "Nathan Cummings", email = "nathan.cummings@ukaea.uk"},
    {name = "James Hodson", email = "james.hodson@ukaea.uk"}
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "dask>=2024.8.0",
    "fastapi-pagination>=0.12.31",
    "fastapi>=0.115.2",
    "jinja2>=3.1.4",
    "numpy<2",
    "pandas==2.1.4",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=17.0.0",
    "sqlakeyset>=2.0.1726021475",
    "sqlalchemy-utils>=0.41.2",
    "sqlalchemy<2",
    "sqlmodel>=0.0.11",
    "strawberry-graphql>=0.246.2",
    "tqdm>=4.66.5",
    "ujson>=5.10.0",
    "uvicorn>=0.32.0",
    "python-keycloak>=5.1.1",
    "python-dotenv>=1.0.1"
]

[project.urls]
repository = "https://github.com/ukaea/fair-mast"

[tool.ruff]
exclude = ["docs/*.ipynb", "notebooks/*.ipynb", "src/metadata/*.ipynb", "src/api/static"]

[tool.uv]
dev-dependencies = [
    "httpx>=0.27.2",
    "pint>=0.24.3",
    "pytest-benchmark>=4.0.0",
    "pytest>=8.3.3",
    "pytest-order>=1.3.0",
    "xarray>=2024.7.0",
    "zarr>=2.18.2",
]